+++ lib/benchmark_driver/runner/ruby_stdout.rb
@@ -5,14 +5,15 @@ require 'shellwords'
 require 'open3'
 
 # Use stdout of ruby command
-class BenchmarkDriver::Runner::CommandStdout
+class BenchmarkDriver::Runner::RubyStdout
   # JobParser returns this, `BenchmarkDriver::Runner.runner_for` searches "*::Job"
   Job = ::BenchmarkDriver::Struct.new(
-    :name,              # @param [String] name - This is mandatory for all runner
-    :command,           # @param [Array<String>]
-    :working_directory, # @param [String,NilClass]
-    :metrics,           # @param [Array<BenchmarkDriver::Metric>]
-    :stdout_to_metrics, # @param [String]
+    :name,                   # @param [String] name - This is mandatory for all runner
+    :command,                # @param [Array<String>]
+    :working_directory,      # @param [String,NilClass]
+    :metrics,                # @param [Array<BenchmarkDriver::Metric>]
+    :value_from_stdout,      # @param [String]
+    :environment_from_stdout # @param [Hash{ String => String }]
   )
   # Dynamically fetched and used by `BenchmarkDriver::JobParser.parse`
   class << JobParser = Module.new
@@ -21,27 +22,43 @@ class BenchmarkDriver::Runner::CommandStdout
     # @param [String,NilClass] working_directory
     # @param [Hash] metrics_type
     # @param [String] stdout_to_metrics
-    def parse(name:, command:, working_directory: nil, metrics_type:, stdout_to_metrics:)
+    def parse(name:, command:, working_directory: nil, metrics:, environment: {})
+      unless metrics.is_a?(Hash)
+        raise ArgumentError.new("metrics must be Hash, but got #{metrics.class}")
+      end
+      if metrics.size == 0
+        raise ArgumentError.new('At least one metric must be specified"')
+      elsif metrics.size != 1
+        raise NotImplementedError.new('Having multiple metrics is not supported yet')
+      end
+
+      metric, value_from_stdout = parse_metric(*metrics.first)
+      environment_from_stdout = Hash[environment.map { |k, v| [k, parse_environment(v)] }]
+
       Job.new(
         name: name,
         command: command.shellsplit,
         working_directory: working_directory,
-        metrics: parse_metrics(metrics_type),
-        stdout_to_metrics: stdout_to_metrics,
+        metrics: [metric],
+        value_from_stdout: value_from_stdout,
+        environment_from_stdout: environment_from_stdout,
       )
     end
 
     private
 
-    def parse_metrics(unit:, name: nil, larger_better: nil, worse_word: nil)
-      name ||= unit
+    def parse_metric(name, unit:, from_stdout:, larger_better: true, worse_word: 'slower')
       metric = BenchmarkDriver::Metric.new(
         name: name,
         unit: unit,
         larger_better: larger_better,
         worse_word: worse_word,
       )
-      [metric]
+      [metric, from_stdout]
+    end
+
+    def parse_environment(from_stdout:)
+      from_stdout
     end
   end
 
@@ -62,17 +79,19 @@ class BenchmarkDriver::Runner::CommandStdout
       jobs.each do |job|
         @output.with_job(name: job.name) do
           @config.executables.each do |exec|
-            best_value = with_repeat(metric) do
+            best_value, environment = with_repeat(metric) do
               stdout = with_chdir(job.working_directory) do
                 with_ruby_prefix(exec) { execute(*exec.command, *job.command) }
               end
-              StdoutToMetrics.new(
+              script = StdoutToMetrics.new(
                 stdout: stdout,
-                stdout_to_metrics: job.stdout_to_metrics,
-              ).metrics_value
+                value_from_stdout: job.value_from_stdout,
+                environment_from_stdout: job.environment_from_stdout,
+              )
+              [script.value, script.environment]
             end
 
-            @output.with_context(name: exec.name, executable: exec) do
+            @output.with_context(name: exec.name, executable: exec, environment: environment) do
               @output.report(value: best_value, metric: metric)
             end
           end
@@ -109,10 +128,10 @@ class BenchmarkDriver::Runner::CommandStdout
 
   # Return multiple times and return the best metrics
   def with_repeat(metric, &block)
-    values = @config.repeat_count.times.map do
+    value_environments = @config.repeat_count.times.map do
       block.call
     end
-    values.sort_by do |value|
+    value_environments.sort_by do |value, _|
       if metric.larger_better
         value
       else
@@ -121,9 +140,17 @@ class BenchmarkDriver::Runner::CommandStdout
     end.last
   end
 
-  StdoutToMetrics = ::BenchmarkDriver::Struct.new(:stdout, :stdout_to_metrics) do
-    def metrics_value
-      eval(stdout_to_metrics, binding)
+  StdoutToMetrics = ::BenchmarkDriver::Struct.new(:stdout, :value_from_stdout, :environment_from_stdout) do
+    def value
+      value = eval(value_from_stdout, binding)
+    end
+
+    def environment
+      ret = {}
+      environment_from_stdout.each do |name, from_stdout|
+        ret[name] = eval(from_stdout, binding)
+      end
+      ret
     end
   end
   private_constant :StdoutToMetrics